+++ R/recipe.R
@@ -12,6 +12,13 @@ recipe <- function(x, ...) {
 #' @rdname recipe
 #' @export
 recipe.default <- function(x, ...) {
+
+  # Doing this here since it should work for all types of Matrix classes
+  if (is_sparse_matrix(x)) {
+    x <- sparsevctrs::coerce_to_sparse_tibble(x)
+    return(recipe(x, ...))
+  }
+
   cli::cli_abort(c(
     x = "{.arg x} should be a data frame, matrix, formula, or tibble.",
     i = "{.arg x} is {.obj_type_friendly {x}}."
@@ -33,7 +40,9 @@ recipe.default <- function(x, ...) {
 #'  Dots are allowed as are simple multivariate outcome terms (i.e. no need for
 #'  `cbind`; see Examples). A model formula may not be the best choice for
 #'  high-dimensional data with many columns, because of problems with memory.
-#' @param x,data A data frame or tibble of the *template* data set
+#' @param x,data A data frame, tibble, or sparse matrix from the `Matrix` 
+#'   package of the *template* data set. See [sparse_data] for more information
+#'   about use of sparse data.
 #'   (see below).
 #' @return An object of class `recipe` with sub-objects:
 #'   \item{var_info}{A tibble containing information about the original data
@@ -208,6 +217,10 @@ recipe.formula <- function(formula, data, ...) {
     ))
   }
 
+  if (is_sparse_matrix(data)) {
+    data <- sparsevctrs::coerce_to_sparse_tibble(data)
+  }
+
   if (!is_tibble(data)) {
     data <- as_tibble(data)
   }
@@ -308,8 +321,9 @@ prep <- function(x, ...) {
 #' For a recipe with at least one preprocessing operation, estimate the required
 #'   parameters from a training set that can be later applied to other data
 #'   sets.
-#' @param training A data frame or tibble that will be used to estimate
-#'   parameters for preprocessing.
+#' @param training A data frame, tibble, or sparse matrix from the `Matrix` 
+#'   package, that will be used to estimate parameters for preprocessing. See 
+#'   [sparse_data] for more information about use of sparse data.
 #' @param fresh A logical indicating whether already trained operation should be
 #'   re-trained. If `TRUE`, you should pass in a data set to the argument
 #'   `training`.
@@ -383,7 +397,7 @@ prep.recipe <-
            log_changes = FALSE,
            strings_as_factors = TRUE,
            ...) {
-    training <- check_training_set(training, x, fresh)
+    training <- validate_training_data(training, x, fresh)
 
     tr_data <- train_info(training)
 
@@ -591,9 +605,11 @@ bake <- function(object, ...) {
 #'   [prep()], apply the computations to new data.
 #' @param object A trained object such as a [recipe()] with at least
 #'   one preprocessing operation.
-#' @param new_data A data frame or tibble for whom the preprocessing will be
-#'   applied. If `NULL` is given to `new_data`, the pre-processed _training
-#'   data_ will be returned (assuming that `prep(retain = TRUE)` was used).
+#' @param new_data A data frame, tibble, or sparse matrix from the `Matrix` 
+#'   package for whom the preprocessing will be applied. If `NULL` is given to 
+#'   `new_data`, the pre-processed _training data_ will be returned (assuming
+#'   that `prep(retain = TRUE)` was used). See [sparse_data] for more 
+#'   information about use of sparse data.
 #' @param ... One or more selector functions to choose which variables will be
 #'   returned by the function. See [selections()] for more details.
 #'   If no selectors are given, the default is to use
@@ -685,6 +701,10 @@ bake.recipe <- function(object, new_data, ..., composition = "tibble") {
     } else {
       cli::cli_abort("Please pass a data set to {.arg new_data}.")
     }
+  }
+
+  if (is_sparse_matrix(new_data)) {
+    new_data <- sparsevctrs::coerce_to_sparse_tibble(new_data)
   }
 
   if (!is_tibble(new_data)) {